+++ src/core/activity-registerer.ts
@@ -1,55 +1,53 @@
-import { BotGuildMember, Client, Logger } from "disharmony";
+import { Client, Logger } from "disharmony";
 import Guild from "../models/guild";
+import GuildMember from "../models/guild-member";
 import Message from "../models/message";
 
 export default class ActivityRegisterer
 {
     public startListening()
     {
-        this.client.onMessage.sub(message => this.registerActivity(message.guild, message.member))
-        this.client.onVoiceStateUpdate.sub(member => this.registerActivity(new Guild(member.djs.guild), member))
+        this.client.onMessage.sub(
+            message => this.registerActivity(message.guild, message.member, message.textChannelName))
+        this.client.onVoiceStateUpdate.sub(
+            args => this.registerActivity(
+                new Guild(args.newMember.djs.guild),
+                args.newMember,
+                args.newMember.voiceChannelName || args.oldMember.voiceChannelName))
     }
 
-    private async registerActivity(guild: Guild, member: BotGuildMember)
+    private async registerActivity(guild: Guild, member: GuildMember, channelName: string)
     {
+        if (!member || !guild || member.id === this.client.botId)
+            return
+
         await guild.loadDocument()
-        if (guild
-            && member && member.id !== this.client.botId
-            && this.isGuildSetUp(guild))
-        {
-            guild.users.set(member.id, new Date())
-            await this.markActiveIfNotIgnored(guild, member, true)
-            await guild.save()
-        }
+        if (!this.isGuildSetUp(guild))
+            return
+
+        if (guild.isMemberIgnored(member))
+            return
+
+        guild.users.set(member.id, new Date())
+        await this.markMemberActive(guild, member, channelName)
+        await guild.save()
     }
 
-    private async markActiveIfNotIgnored(guild: Guild, member: BotGuildMember, isActive: boolean)
+    private async markMemberActive(guild: Guild, member: GuildMember, channelName: string)
     {
-        const addRole = isActive ? guild.activeRoleId : guild.inactiveRoleId
-        const removeRole = isActive ? guild.inactiveRoleId : guild.activeRoleId
-
         try
         {
-            if (this.isMemberIgnored(guild, member))
-                return
+            const reasonStr = `Activity detected in channel '${channelName}'`
+            await member.addRole(guild.activeRoleId, reasonStr);
 
-            if (addRole && addRole !== "disabled")
-                await member.addRole(addRole);
-
-            if (removeRole && removeRole !== "disabled")
-                await member.removeRole(removeRole)
+            if (guild.inactiveRoleId && guild.inactiveRoleId !== "disabled")
+                await member.removeRole(guild.inactiveRoleId, reasonStr)
         }
         catch (e)
         {
-            Logger.debugLogError(`Error marking user ${member.username} ${isActive ? "active" : "inactive"} in guild ${guild.name}.`, e)
+            Logger.debugLogError(`Error marking user ${member.username} active in guild ${guild.name}.`, e)
+            Logger.logEvent("ErrorMarkingActive", { guildId: guild.id, memberName: member.username })
         }
-    }
-
-    private isMemberIgnored(guild: Guild, member: BotGuildMember)
-    {
-        const isIgnoredUser = guild.ignoredUserIds.indexOf(member.id) >= 0
-        const hasIgnoredRole = guild.ignoredRoleIds.some(roleId => member.hasRole(roleId))
-        return isIgnoredUser || hasIgnoredRole
     }
 
     private isGuildSetUp(guild: Guild)
@@ -58,6 +56,6 @@ export default class ActivityRegisterer
     }
 
     constructor(
-        private client: Client<Message>,
+        private client: Client<Message, GuildMember>,
     ) { }
 }