+++ src/core/inactivity-manager.ts
@@ -1,5 +1,6 @@
 import { LightClient, loadConfig, Logger } from "disharmony";
 import Guild from "../models/guild";
+import GuildMember from "../models/guild-member";
 
 export default class InactivityManager
 {
@@ -24,36 +25,48 @@ export default class InactivityManager
             return
 
         Logger.debugLog(`Managing inactives for guild ${guild.name}`)
-        for (const member of guild.activeRole.members.values())
+        for (const djsMember of guild.activeRole.members.values())
         {
-            // don't ask me why, sometimes member is null
+            const member = new GuildMember(djsMember)
+            // Don't ask me why, sometimes member is null
             if (!member)
-                return
+                continue
 
+            if (guild.isMemberIgnored(member))
+                continue
+
+            // If the member has the active role but isn't in the database, add them
             const now = new Date()
             if (!guild.users.get(member.id))
             {
-                guild.users.set(member.id, now) // if a user has the active role but isn't in the database, add them
-                Logger.debugLog(`User ${member.user.username} has active role but not found in database, adding new entry`)
+                guild.users.set(member.id, now)
+                Logger.debugLog(`User ${member.username} has active role but not found in database, adding new entry`)
+                Logger.logEvent("FoundManuallyActiveMember", { guildId: guild.id })
+                continue
             }
-            else if (this.isInactiveBeyondThreshold(guild.users.get(member.id)!, now, guild.inactiveThresholdDays))
-            {
-                try
-                {
-                    await member.removeRole(guild.activeRole)
-                    if (guild.inactiveRoleId && guild.inactiveRoleId !== "disabled")
-                        await member.addRole(guild.inactiveRoleId)
-                    guild.users.delete(member.id)
-                    Logger.debugLog(`Updated now inactive user ${member.user.username}`)
-                }
-                catch (e)
-                {
-                    Logger.debugLogError(`Error switching user ${member.user.username} to inactive in guild ${guild.name}`, e)
-                }
-            }
+
+            // If the member is inactive, remove the active role and remove them from the database
+            const isMemberInactive = this.isInactiveBeyondThreshold(guild.users.get(member.id)!, now, guild.inactiveThresholdDays)
+            if (isMemberInactive)
+                await this.markMemberInactive(guild, member)
+                    .catch(e =>
+                    {
+                        Logger.debugLogError(`Error switching user ${member.username} to inactive in guild ${guild.name}`, e)
+                        Logger.logEvent("ErrorMarkingInactive", { guildId: guild.id, memberName: member.username })
+                    })
         }
 
         await guild.save()
+    }
+
+    private async markMemberInactive(guild: Guild, member: GuildMember)
+    {
+        const reasonStr = `No activity detected within last ${guild.inactiveThresholdDays} days`
+        await member.removeRole(guild.activeRole!, reasonStr)
+        if (guild.inactiveRoleId && guild.inactiveRoleId !== "disabled")
+            await member.addRole(guild.inactiveRoleId, reasonStr)
+        guild.users.delete(member.id)
+        Logger.logEvent("MarkedMemberInactive", { memberName: member.username })
     }
 
     private isInactiveBeyondThreshold(lastActiveDate: Date, now: Date, thresholdDays: number): boolean
@@ -73,7 +86,7 @@ if (!module.parent)
     const { config } = loadConfig(configPath)
     const client = new LightClient(config)
     const inactivityManager = new InactivityManager(client)
-    client.initialize(config.token)
+    client.login(config.token)
         .then(async () =>
         {
             await inactivityManager.manageInactiveUsersInAllGuilds()
@@ -84,6 +97,7 @@ if (!module.parent)
         .catch(async err =>
         {
             await Logger.debugLogError("Error running the inactivity monitor", err)
+            await Logger.logEvent("ErrorStartingInactivityMonitor")
             process.exit(1)
         })
 }